Add image and selected image to segment at index.
Modify a segment by changing it's title, adding an image and selected image for a given index.

This is more of a proof of concept as it really only works when all other properties are set. It isn't backwards compatible. For example, if you change the titleFont after setting a segments images it would reset the segment to plain text.